+++ spec/honey_format/configuration_spec.rb
@@ -5,7 +5,7 @@ require 'spec_helper'
 RSpec.describe HoneyFormat::Configuration do
   describe '#header_converter=' do
     it 'can set header converter from Symbol' do
-      expected = HoneyFormat.value_converter[:upcase]
+      expected = HoneyFormat.converter[:upcase]
 
       config = described_class.new
       config.header_converter = :upcase
@@ -14,7 +14,7 @@ RSpec.describe HoneyFormat::Configuration do
     end
 
     it 'can set header converter' do
-      expected = HoneyFormat.value_converter[:upcase]
+      expected = HoneyFormat.converter[:upcase]
 
       config = described_class.new
       config.header_converter = expected
@@ -24,8 +24,8 @@ RSpec.describe HoneyFormat::Configuration do
   end
 
   describe '#converter' do
-    it 'returns a value converter' do
-      expect(described_class.new.converter).to be_a(HoneyFormat::ValueConverter)
+    it 'returns a converter registry' do
+      expect(described_class.new.converter).to be_a(HoneyFormat::ConverterRegistry)
     end
   end
 end