+++ spec/honey_format/converters/header_column_converter_spec.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
 require 'spec_helper'
-require 'honey_format/header_column_converter'
+require 'honey_format/converters/header_column_converter'
 
 describe HoneyFormat::HeaderColumnConverter do
   # See https://bugs.ruby-lang.org/issues/10085
-  ruby_version_under_2_4 = Gem::Version.new(RUBY_VERSION) < Gem::Version.new('2.4.0') # rubocop:disable Naming/VariableNumber
+  affected_ruby_version = Gem::Version.new(RUBY_VERSION) < Gem::Version.new('2.4.0')
   [
     ['first_name', :first_name],
     ['first name', :first_name],
@@ -23,7 +23,7 @@ describe HoneyFormat::HeaderColumnConverter do
     ['first-name', :first_name],
     ['USeRnaMe', :username],
     [nil, :column3, 3],
-    ruby_version_under_2_4 ? ['ÅÄÖ', :'ÅÄÖ'] : ['ÅÄÖ', :'åäö']
+    affected_ruby_version ? ['ÅÄÖ', :'ÅÄÖ'] : ['ÅÄÖ', :'åäö']
   ].each do |data|
     input, expected, index = data